Fix negative LSNs

+++ include/aggregate.h
@@ -10,7 +10,7 @@ public:
   void write(std::ostream& out, std::vector<File>& records);
   void update_records(std::vector<File>& records);
 
-  int record, par_record, previous_par_record, usn_lsn, type, source;
+  unsigned long long record, par_record, previous_par_record, usn_lsn, type, source;
   std::string timestamp, file_name, previous_file_name;
 };
